Navbar + margin styling
Navbar is done but could benefit from an accessibility review at some point (medium priority) i.e. semantics and implementation of aria attributes (also I don't think its the right colour, I couldn't find the overlay colour in variables). The Navbar, radial gradient, and content wrapper styles were all added to the layout.tsx for consistency across pages. The content wrapper should be reviewed. The margin top applied to it is due to the navbar being fixed and as a result it is not in the regular document flow so we need something to push body content below the nav. @christinaraganit you might have a better approach to this. Also moved the styleguide page to its own directory so it doesn't conflict with the default landing page when we merge branches. footer was changed to self closing tag as react documentation says that you only need to use a closing tag if your component has children. My bad for long commit.