52 about hero section (#53)
Could not find the pull request template as mentioned in the notion.
Hero section completed. It contains none of the proper styling from our
styling guide (all just placeholder for the time being). The styling on
the <main> element should be added to the layout.tsx or classed into the
styling branch as a preset to make for easier page setup.
@ChristopherFong604

---------

LGTM

Co-authored-by: ChrisFong604 <ccffoonngg@gmail.com>