68 Posthog Integration (#75)
## Describe your changes
This PR adds the default setup for posthog to the website. It also adds
to the lib some types for mouse activity and a posthog event wrapper
function, but is not used in the code itself until necessary.

Some extraneous console log commands were removed as they are no longer
needed

Also fixed my name 🤡 

## Testing steps 
1. checkout to 68-posthog
2. run `pnpm install` 
3. launch the application, and open DevTools
4. Note that on page load, and whenever a link or button is clicked, a
POST request is sent to the posthog collection server
5. Login to [posthog](https://us.posthog.com) with `sfusurge@gmail.com`.
The credentials can be found in the notion
6. Open the Product Analytics and/or Web Analytics tab, and browse
through the data. You should hopefully be able to see your own personal
activity be recorded and persisted in the dashboard, although it may
take some time to register

## Issue ticket number and link
https://github.com/sfusurge/surge_website/issues/68

## Checklist before requesting a review
- [ ] I have performed a self-review of my code
- [ ] If it is a core feature, I have added thorough tests.
- [ ] Do we need to implement analytics?